feat: Notations for Java packages and constants
fix: also handle Java keywords
