+++ src/test/kotlin/NotationsTest.kt
@@ -1,56 +1,5 @@
 package de.joshuagleitze.stringnotation
 
-import ch.tutteli.atrium.api.fluent.en_GB.feature
-import ch.tutteli.atrium.api.fluent.en_GB.toBe
-import ch.tutteli.atrium.api.verbs.expect
-import org.junit.jupiter.api.TestInstance
-import org.junit.jupiter.api.TestInstance.Lifecycle.PER_CLASS
-import org.junit.jupiter.params.ParameterizedTest
-import org.junit.jupiter.params.provider.Arguments.arguments
-import org.junit.jupiter.params.provider.MethodSource
-
-@TestInstance(PER_CLASS)
-abstract class BaseNotationTest(
-	private val notation: StringNotation,
-	private val unchangedWords: List<Pair<String, Word>>,
-	private val parseOnlyWords: List<Pair<String, Word>> = emptyList(),
-	private val printOnlyWords: List<Pair<Word, String>> = emptyList()
-) {
-	@ParameterizedTest(name = "\"{0}\" -> {1}")
-	@MethodSource("parseWords")
-	fun `parses words correctly`(input: String, expectedWord: Word) {
-		expect(input.fromNotation(notation)) {
-			feature(Word::partsList).toBe(expectedWord.partsList)
-		}
-	}
-
-	@ParameterizedTest(name = "{1} -> \"{0}\"")
-	@MethodSource("printWords")
-	fun `prints words correctly`(sourceWord: Word, expectedResult: String) {
-		expect(sourceWord) {
-			feature(Word::toNotation, notation).toBe(expectedResult)
-		}
-	}
-
-	@ParameterizedTest(name = "\"{0}\"")
-	@MethodSource("unchangedWords")
-	fun `parsing and printing a word written in this notation does not change the word`(word: String) {
-		expect(word) {
-			feature(String::fromNotation, notation) {
-				feature(Word::toNotation, notation).toBe(word)
-			}
-		}
-	}
-
-	private fun parseWords() = asArguments(unchangedWords + parseOnlyWords)
-	private fun printWords() = asArguments(unchangedWords.map { it.swap() } + printOnlyWords)
-	private fun unchangedWords() = asArguments(unchangedWords)
-
-	private fun asArguments(pairs: List<Pair<*, *>>) = pairs.map { arguments(it.first, it.second) }
-
-	private fun <First, Second> Pair<First, Second>.swap() = Pair(this.second, this.first)
-}
-
 class UpperCamelCaseTest: BaseNotationTest(
 	notation = UpperCamelCase,
 	unchangedWords = listOf("ImInUpperCamelCase" to Word("im", "in", "upper", "camel", "case")),
@@ -70,7 +19,8 @@ class LowerCamelCaseTest: BaseNotationTest(
 class ScreamingSnakeCaseTest: BaseNotationTest(
 	notation = ScreamingSnakeCase,
 	unchangedWords = listOf("IM_IN_SCREAMING_SNAKE_CASE" to Word("im", "in", "screaming", "snake", "case")),
-	parseOnlyWords = listOf("im_iN_sNAKe_cASE_with_CAPItals" to Word("im", "in", "snake", "case", "with", "capitals"))
+	parseOnlyWords = listOf("im_iN_sNAKe_cASE_with_CAPItals" to Word("im", "in", "snake", "case", "with", "capitals")),
+	printOnlyWords = listOf(Word("im", "iN", "sNAKe", "cASE", "with", "CAPItals") to "IM_IN_SNAKE_CASE_WITH_CAPITALS")
 )
 
 class SnakeCaseTest: BaseNotationTest(
@@ -78,29 +28,6 @@ class SnakeCaseTest: BaseNotationTest(
 	unchangedWords = listOf(
 		"im_in_snake_case" to Word("im", "in", "snake", "case"),
 		"im_iN_sNAKe_cASE_with_CAPItals" to Word("im", "iN", "sNAKe", "cASE", "with", "CAPItals")
-	)
-)
-
-class JavaTypeNameTest: BaseNotationTest(
-	notation = JavaTypeName,
-	unchangedWords = listOf("ImInJavaTypeNotation" to Word("im", "in", "java", "type", "notation")),
-	printOnlyWords = listOf(
-		Word("I’m using", "Bad", "chaRacters!") to "ImusingBadCharacters",
-		Word("1", "type", "name", "4", "you") to "TypeName4You",
-		Word("removes", "upperCase") to "RemovesUppercase"
-	)
-)
-
-class JavaMemberNameTest: BaseNotationTest(
-	notation = JavaMemberName,
-	unchangedWords = listOf("imInJavaMemberNotation" to Word("im", "in", "java", "member", "notation")),
-	printOnlyWords = listOf(
-		Word("I’m using", "Bad", "chaRacters!") to "imusingBadCharacters",
-		Word("1", "Member", "name", "4", "you") to "memberName4You",
-		Word("_", "underscore", "start") to "_underscoreStart",
-		Word("$", "dollar", "start") to "\$dollarStart",
-		Word("a", "letter", "start") to "aLetterStart",
-		Word("removes", "upperCase") to "removesUppercase"
 	)
 )
